pyproject.toml
[project]
name = "fastapi-grpc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0",
    "uvicorn>=0.35.0",
]
